implement hover anchor for quad glyphs (#9013)
* copied anchor point function from quad.ts

* removed anchor point function from quad.ts

* more natural ordering of anchor points
