+++ bokehjs/src/lib/models/glyphs/quad.ts
@@ -1,6 +1,5 @@
 import {Box, BoxView, BoxData} from "./box"
 import {Arrayable} from "core/types"
-import {Anchor} from "core/enums"
 import {SpatialIndex} from "core/util/spatial"
 import * as p from "core/properties"
 
@@ -21,26 +20,6 @@ export interface QuadView extends QuadData {}
 export class QuadView extends BoxView {
   model: Quad
   visuals: Quad.Visuals
-
-  get_anchor_point(anchor: Anchor, i: number, _spt: [number, number]): {x: number, y: number} | null {
-    const left = Math.min(this.sleft[i], this.sright[i])
-    const right = Math.max(this.sright[i], this.sleft[i])
-    const top = Math.min(this.stop[i], this.sbottom[i])     // screen coordinates !!!
-    const bottom = Math.max(this.sbottom[i], this.stop[i])  //
-
-    switch (anchor) {
-      case "top_left":      return {x: left,             y: top             }
-      case "top_center":    return {x: (left + right)/2, y: top             }
-      case "top_right":     return {x: right,            y: top             }
-      case "center_right":  return {x: right,            y: (top + bottom)/2}
-      case "bottom_right":  return {x: right,            y: bottom          }
-      case "bottom_center": return {x: (left + right)/2, y: bottom          }
-      case "bottom_left":   return {x: left,             y: bottom          }
-      case "center_left":   return {x: left,             y: (top + bottom)/2}
-      case "center":        return {x: (left + right)/2, y: (top + bottom)/2}
-      default:              return null
-    }
-  }
 
   scenterx(i: number): number {
     return (this.sleft[i] + this.sright[i])/2
